move Metricsmanager
